chore: :truck: Move data related hooks to their own folder
